Simplify the Thumb attributes
Having 4 attributes for the Thumb, and keeping the image around in
separate Bitmaps, is slightly wasteful. And relying on Bitmap for
the thumbs means that it's impossible to use any non-Bitmap drawable
for the thumb resources (i.e., a <shape/> oval drawable).

So by allowing for any Drawable to be used (including a BitmapDrawable),
we not only allow shape and color drawables to be used, but we also
allow those drawables to now be stateful (e.g., see the new
`res/drawable/seek_thumb.xml`)

If you previously relied upon the `thumbImageNormal`/-`Pressed`
attributes, you should convert that into a `StateListDrawable`, the
way it was done here in `res/drawable/seek_thumb.xml`.

If you are using the `thumbColorNormal`/`thumbColorPressed` behavior,
you should also switch that to a `StateListDrawable`, where the
`state_pressed` item references an oval shape filled with your desired
color.  These changes represent the "Android Way" of using drawables.