Fix some px/dp scaling bugs
Resolving dimension attributes already applies the correct density
scaling, so re-applying the scale to a value that was already specified
using DIP is wrong.

Instead, initialize the default attribute values to -1, and allow the
default to be scaled only if it was not specified (and scaled) already.

Also remove anti-aliasing flag from the bar paints, because it serves
no purpose on a line.