Allow sloppy seek gestures
If the touch even is not directly within the Thumb's target area, but
it is within the "slop" range, treat it as a DOWN event on that Thumb.

Otherwise, if the down event is outside of the slop range, it will be
treated as a touch-to-seek event when the UP event happens.