Also pass the X-offset of the left and right thumbs to the index-changed listener.
This allows the listener to perform a visual feedback update to track
the index changes.  E.g., to show a popup above or below the Thumb,
or move a textview containing the left and right values, so that they
are adjacent to their respective thumbs, etc.

If not needed, those callback arguments can be ignored.